Math-Change/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
supabase>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
httpx>=0.24.0
boto3>=1.28.0
python-multipart>=0.0.6
